requirements.txt
asgiref==3.5.2
Django==4.0.5
sqlparse==0.4.2
tzdata==2022.1
uWSGI>=2.0.18,<2.1
django-mysql==3.8.1
mysqlclient==2.1.1
APScheduler==3.9.1
sqlparse==0.4.2
requests==2.28.0
django-environ==0.9.0
python-dotenv==0.19.0
debugpy>=1.6.2,<1.7
djangorestframework==3.13.1
drf-spectacular>=0.22.1,<0.23